Renaming for pytest, and adding more specific unit tests.
